added botadmin+nfc refactored
